+++ modelchecking/m3c/src/main/java/module-info.java
@@ -45,6 +45,7 @@ open module net.automatalib.modelchecker.m3c {
     // annotations are 'provided'-scoped and do not need to be loaded at runtime
     requires static org.checkerframework.checker.qual;
 
+    exports net.automatalib.modelchecker.m3c.checker;
     exports net.automatalib.modelchecker.m3c.formula;
     exports net.automatalib.modelchecker.m3c.formula.ctl;
     exports net.automatalib.modelchecker.m3c.formula.modalmu;