+++ modelchecking/m3c/src/main/java/net/automatalib/modelchecker/m3c/checker/GenericTypedChecker.java
@@ -1,0 +1,41 @@
+/* Copyright (C) 2013-2026 TU Dortmund University
+ * This file is part of AutomataLib <https://automatalib.net>.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package net.automatalib.modelchecker.m3c.checker;
+
+import java.util.Collection;
+import java.util.function.Function;
+
+import net.automatalib.graph.ContextFreeModalProcessSystem;
+import net.automatalib.modelchecker.m3c.formula.FormulaNode;
+import net.automatalib.modelchecker.m3c.solver.AbstractDDSolver;
+import net.automatalib.modelchecker.m3c.solver.WitnessTree;
+import net.automatalib.modelchecking.ModelChecker;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+class GenericTypedChecker<L, AP> extends InternalChecker<L, AP>
+        implements ModelChecker<L, ContextFreeModalProcessSystem<L, AP>, FormulaNode<L, AP>, WitnessTree<L, AP>> {
+
+    GenericTypedChecker(Function<? super ContextFreeModalProcessSystem<L, AP>, ? extends AbstractDDSolver<?, L, AP>> creator) {
+        super(creator);
+    }
+
+    @Override
+    public @Nullable WitnessTree<L, AP> findCounterExample(ContextFreeModalProcessSystem<L, AP> cfmps,
+                                                           Collection<? extends L> inputs,
+                                                           FormulaNode<L, AP> property) {
+        return doFindCounterExample(cfmps, property);
+    }
+}